Added bench for many gens groups
